feat: update BuntalConfig type and configuration handling in bundler and runServer
