+++ packages/buntal/index.ts
@@ -1,3 +1,10 @@
 export type { Req, Res } from '@buntal/core'
-export type { BuntalConfig } from 'buntal/server'
 export * from './components'
+
+export type BuntalConfig = {
+  env?: 'development' | 'production'
+  appDir?: string
+  outDir?: string
+  staticDir?: string
+  config?: Partial<Bun.BuildConfig>
+}
